feat(data-debruijn): support snat-as-word8, ix-as-word8, and th-as-bitvec

+++ src-arbitrary/Data/DeBruijn/Index/Fast/Arbitrary.hs
@@ -19,7 +19,7 @@ instance Arbitrary SomeIx where
   arbitrary = do
     Positive boundOverIndex <- arbitrary
     NonNegative index <- arbitrary
-    pure $ toSomeIxRaw (index + boundOverIndex, index)
+    pure $ toSomeIxRaw (fromIntegral index + boundOverIndex, index)
 
 instance Arbitrary (Ix (S Z)) where
   arbitrary :: Gen (Ix (S Z))
